fixed plot imports after autoarray refactor
